Revert "raft,raftgorums: Make catchupDiff uint64"
This reverts commit be73b0a1e9187657e5530c85bbc3c9d1d86e26ca.

+++ raftgorums/raft.go
@@ -120,7 +120,7 @@ type Raft struct {
 
 	catchingup   time.Time
 	catchupIndex uint64
-	catchupDiff  uint64
+	catchupDiff  int
 }
 
 func (r *Raft) incCmd() {
